requirements.txt
psycopg2-binary
pydantic
numpy
scikit-learn
openai
python-dotenv
termcolor
PyYAML
redis